Mention requiring the gem in the README
I'm seeing this error:

    unable to connect to chromedriver 127.0.0.1:9515 (Selenium::WebDriver::Error::WebDriverError)

PR #58 changes the "chromedriver" executable file name to "chromedriver-helper", but I found that Selenium::WebDriver::Chrome.driver_path was nil after trying to visit a page (after rescuing from the failure). 

The problem is that I was never loading the new chromedriver-helper code to set the custom driver_path.  I added this to features/support/env.rb and my Cucumber scenario started working:

    require 'chromedriver-helper'

I'm using this outside of Rails in a standalone test suite to test an app that is developed separately.  (Rails may be requiring the gem behind the scenes.)

Mention the possible need to require 'chromedriver-helper' in the Standalone section of the README, where people will be looking for installation instructions.